refactor: split colonizethis_models monoliths + add models file-size gate
Phase 5 of #3393. Splits the three largest hand-written model files
(app_events.dart, world_market.dart, orders.dart) into concern-grouped
`part` files, each below 500 non-comment lines, preserving the public
colonizethis_models barrel contract. Adds a dedicated repo lint rule
`repo.models_file_size` (500 non-comment line cap for
packages/colonizethis_models/lib/src) with a passing check suite; game.dart
is a documented grandfathered baseline pending a follow-up split.

Refs #3393
