+++ packages/colonizethis_models/lib/src/orders.dart
@@ -4,6 +4,9 @@ import 'province_id.dart';
 import 'worker_tier.dart';
 import 'world_market.dart';
 
+part 'orders/move_orders.dart';
+part 'orders/build_work_orders.dart';
+
 /// Per-player orders for the current turn.
 /// SPEC/game/world-model.
 /// SPEC/program/orders.md
@@ -394,400 +397,4 @@ class Orders {
     }
     return true;
   }
-}
-
-/// Move a civilian land unit to a destination **land tile**.
-/// Serialized form is [destinationTileKey] only; enclosing province is derived.
-/// SPEC/program/orders.md; SPEC/game (issue #1877 civilian tile moves).
-class MoveOrder {
-  const MoveOrder({required this.unitId, required this.destinationTileKey});
-
-  final String unitId;
-
-  /// Land tile key `regionId|localProvinceId|x|y`. Province id is derived from the first two segments.
-  final String destinationTileKey;
-
-  Map<String, dynamic> toJson() => {
-    'unitId': unitId,
-    'destinationTileKey': destinationTileKey,
-  };
-
-  static MoveOrder fromJson(Map<String, dynamic> json) {
-    final unitId = json['unitId'] as String;
-    final tile = json['destinationTileKey'] as String?;
-    if (tile != null && tile.isNotEmpty) {
-      return MoveOrder(unitId: unitId, destinationTileKey: tile);
-    }
-    final legacy = json['destinationProvinceId'] as String?;
-    if (legacy != null && legacy.isNotEmpty) {
-      if (!ProvinceId.isPrefixed(legacy)) {
-        throw ModelValidationException(
-          'MoveOrder legacy destinationProvinceId must be prefixed (regionId|localId): "$legacy"',
-        );
-      }
-      final region = ProvinceId.regionIdFrom(legacy);
-      final local = ProvinceId.localIdFrom(legacy);
-      return MoveOrder(
-        unitId: unitId,
-        destinationTileKey: '$region|$local|0|0',
-      );
-    }
-    throw ModelValidationException(
-      'MoveOrder requires destinationTileKey (or legacy destinationProvinceId)',
-    );
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is MoveOrder &&
-          runtimeType == other.runtimeType &&
-          unitId == other.unitId &&
-          destinationTileKey == other.destinationTileKey;
-
-  @override
-  int get hashCode => Object.hash(unitId, destinationTileKey);
-}
-
-/// Move an army (all its regiments) to a province. SPEC/game/military-armies.md.
-class ArmyMoveOrder {
-  const ArmyMoveOrder({
-    required this.armyId,
-    required this.destinationProvinceId,
-  });
-
-  final String armyId;
-  final String destinationProvinceId;
-
-  Map<String, dynamic> toJson() => {
-    'armyId': armyId,
-    'destinationProvinceId': destinationProvinceId,
-  };
-
-  static ArmyMoveOrder fromJson(Map<String, dynamic> json) {
-    final destinationProvinceId = ProvinceId.requirePrefixed(
-      json['destinationProvinceId'] as String,
-      fieldName: 'ArmyMoveOrder.destinationProvinceId',
-    );
-    return ArmyMoveOrder(
-      armyId: json['armyId'] as String,
-      destinationProvinceId: destinationProvinceId,
-    );
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is ArmyMoveOrder &&
-          runtimeType == other.runtimeType &&
-          armyId == other.armyId &&
-          destinationProvinceId == other.destinationProvinceId;
-
-  @override
-  int get hashCode => Object.hash(armyId, destinationProvinceId);
-}
-
-/// Move a fleet to an adjacent sea zone or dock at a port. Phase 5. SPEC/program/naval-movement-resolution.md.
-///
-/// Exactly one of [destinationSeaZoneId] (move to sea) or [destinationPortProvinceId] (dock)
-/// must be set. Backward compat: fromJson accepts legacy payloads with only destinationSeaZoneId.
-class NavalMoveOrder {
-  /// Exactly one of [destinationSeaZoneId] or [destinationPortProvinceId] must be set (enforced in [fromJson] and validation).
-  const NavalMoveOrder({
-    required this.fleetId,
-    this.destinationSeaZoneId,
-    this.destinationPortProvinceId,
-  });
-
-  final String fleetId;
-
-  /// Non-null for "move to sea zone". Null when [destinationPortProvinceId] is set (dock).
-  final String? destinationSeaZoneId;
-
-  /// Non-null for "dock at province". Null when [destinationSeaZoneId] is set.
-  final String? destinationPortProvinceId;
-
-  /// True when this order is a dock (destination is a port province).
-  bool get isDock =>
-      destinationPortProvinceId != null &&
-      destinationPortProvinceId!.isNotEmpty;
-
-  Map<String, dynamic> toJson() => {
-    'fleetId': fleetId,
-    if (destinationSeaZoneId != null)
-      'destinationSeaZoneId': destinationSeaZoneId,
-    if (destinationPortProvinceId != null)
-      'destinationPortProvinceId': destinationPortProvinceId,
-  };
-
-  static NavalMoveOrder fromJson(Map<String, dynamic> json) {
-    final portId = json['destinationPortProvinceId'] as String?;
-    final seaId = json['destinationSeaZoneId'] as String?;
-    final isDock = portId != null && portId.isNotEmpty;
-    final String? destSea;
-    final String? destPort;
-    if (isDock) {
-      destSea = null;
-      destPort = portId;
-    } else {
-      if (seaId == null || seaId.isEmpty) {
-        throw ModelValidationException(
-          'destinationSeaZoneId required for move to sea',
-        );
-      }
-      destSea = seaId;
-      destPort = null;
-    }
-    return NavalMoveOrder(
-      fleetId: json['fleetId'] as String,
-      destinationSeaZoneId: destSea,
-      destinationPortProvinceId: destPort,
-    );
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is NavalMoveOrder &&
-          runtimeType == other.runtimeType &&
-          fleetId == other.fleetId &&
-          destinationSeaZoneId == other.destinationSeaZoneId &&
-          destinationPortProvinceId == other.destinationPortProvinceId;
-
-  @override
-  int get hashCode =>
-      Object.hash(fleetId, destinationSeaZoneId, destinationPortProvinceId);
-}
-
-/// Assign a mission to a fleet (patrol, blockade, beachhead, defend). SPEC/program/naval-movement-resolution.md.
-class NavalMissionOrder {
-  const NavalMissionOrder({
-    required this.fleetId,
-    required this.mission,
-    this.targetPortId,
-    this.targetProvinceId,
-  });
-
-  final String fleetId;
-  final String mission;
-  final String? targetPortId;
-  final String? targetProvinceId;
-
-  Map<String, dynamic> toJson() => {
-    'fleetId': fleetId,
-    'mission': mission,
-    if (targetPortId != null) 'targetPortId': targetPortId,
-    if (targetProvinceId != null) 'targetProvinceId': targetProvinceId,
-  };
-
-  static NavalMissionOrder fromJson(Map<String, dynamic> json) {
-    return NavalMissionOrder(
-      fleetId: json['fleetId'] as String,
-      mission: json['mission'] as String? ?? 'none',
-      targetPortId: json['targetPortId'] as String?,
-      targetProvinceId: json['targetProvinceId'] as String?,
-    );
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is NavalMissionOrder &&
-          runtimeType == other.runtimeType &&
-          fleetId == other.fleetId &&
-          mission == other.mission &&
-          targetPortId == other.targetPortId &&
-          targetProvinceId == other.targetProvinceId;
-
-  @override
-  int get hashCode =>
-      Object.hash(fleetId, mission, targetPortId, targetProvinceId);
-}
-
-/// Build a new unit for a player.
-/// SPEC/program/orders.md
-class BuildUnitOrder {
-  const BuildUnitOrder({
-    required this.unitType,
-    required this.isMilitary,
-    required this.spawnProvinceId,
-  });
-
-  final String unitType;
-  final bool isMilitary;
-  final String spawnProvinceId;
-
-  Map<String, dynamic> toJson() => {
-    'unitType': unitType,
-    'isMilitary': isMilitary,
-    'spawnProvinceId': spawnProvinceId,
-  };
-
-  static BuildUnitOrder fromJson(Map<String, dynamic> json) {
-    final spawnProvinceId = ProvinceId.requirePrefixed(
-      json['spawnProvinceId'] as String,
-      fieldName: 'BuildUnitOrder.spawnProvinceId',
-    );
-    return BuildUnitOrder(
-      unitType: json['unitType'] as String,
-      isMilitary: json['isMilitary'] as bool? ?? false,
-      spawnProvinceId: spawnProvinceId,
-    );
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is BuildUnitOrder &&
-          runtimeType == other.runtimeType &&
-          unitType == other.unitType &&
-          isMilitary == other.isMilitary &&
-          spawnProvinceId == other.spawnProvinceId;
-
-  @override
-  int get hashCode => Object.hash(unitType, isMilitary, spawnProvinceId);
-}
-
-/// Order a unit to work (explore, build improvement, prospect).
-/// Applies to civilian units only. SPEC/program/orders.md
-class WorkOrder {
-  const WorkOrder({
-    required this.unitId,
-    required this.target,
-    required this.targetTileKey,
-  });
-
-  final String unitId;
-  final String target;
-
-  /// Tile key for the work target (format regionId|provinceId|x|y). For province-level actions (e.g. explore) a synthetic key may be used.
-  final String targetTileKey;
-
-  Map<String, dynamic> toJson() => {
-    'unitId': unitId,
-    'target': target,
-    'targetTileKey': targetTileKey,
-  };
-
-  static WorkOrder fromJson(Map<String, dynamic> json) {
-    return WorkOrder(
-      unitId: json['unitId'] as String,
-      target: json['target'] as String,
-      targetTileKey: json['targetTileKey'] as String? ?? '',
-    );
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is WorkOrder &&
-          runtimeType == other.runtimeType &&
-          unitId == other.unitId &&
-          target == other.target &&
-          targetTileKey == other.targetTileKey;
-
-  @override
-  int get hashCode => Object.hash(unitId, target, targetTileKey);
-}
-
-/// Funding level for research per slot. Maps to treasury cost and research
-/// points per turn. SPEC/program/research-resolution.md
-enum ResearchFundingLevel { none, low, medium, high, maximum }
-
-/// Per-slot research assignment. Phase 5.
-class ResearchOrder {
-  const ResearchOrder({
-    required this.slotIndex,
-    required this.techId,
-    required this.funding,
-  });
-
-  /// Zero-based slot index (0..N-1).
-  final int slotIndex;
-
-  /// Tech id to research in this slot. Empty string = cancel slot.
-  final String techId;
-
-  /// Funding level for this slot.
-  final ResearchFundingLevel funding;
-
-  Map<String, dynamic> toJson() => {
-    'slotIndex': slotIndex,
-    'techId': techId,
-    'funding': funding.name,
-  };
-
-  static ResearchOrder fromJson(Map<String, dynamic> json) {
-    final fundingRaw =
-        json['funding'] as String? ?? ResearchFundingLevel.none.name;
-    final funding = ResearchFundingLevel.values.firstWhere(
-      (e) => e.name == fundingRaw,
-      orElse: () => ResearchFundingLevel.none,
-    );
-    return ResearchOrder(
-      slotIndex: (json['slotIndex'] as num).toInt(),
-      techId: json['techId'] as String? ?? '',
-      funding: funding,
-    );
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is ResearchOrder &&
-          runtimeType == other.runtimeType &&
-          slotIndex == other.slotIndex &&
-          techId == other.techId &&
-          funding == other.funding;
-
-  @override
-  int get hashCode => Object.hash(slotIndex, techId, funding);
-}
-
-/// Recruit or train a worker into the player's WorkerPool.
-///
-/// One queued order increments the target tier by one. Non-peasant tiers
-/// additionally consume one peasant (the "train" mapping per SPEC); UI may
-/// expose distinct Recruit and Train controls per tier but both emit this
-/// order type.
-///
-/// Resolved in Build / work (phase 12) before [BuildUnitOrder] per
-/// SPEC/program/turn-resolution-phase-details.md § Build / work.
-///
-/// SPEC/game/workers-and-population.md § Recruiting, Training, and
-/// Disbanding (authoritative cost table and rejection vocabulary).
-/// SPEC/program/orders.md § Order Types.
-class RecruitWorkerOrder {
-  const RecruitWorkerOrder({required this.targetTier});
-
-  /// Worker tier to add to the pool.
-  final WorkerTier targetTier;
-
-  Map<String, dynamic> toJson() => {'targetTier': targetTier.id};
-
-  static RecruitWorkerOrder fromJson(Map<String, dynamic> json) {
-    final raw = json['targetTier'];
-    if (raw is! String || raw.isEmpty) {
-      throw ModelValidationException(
-        'RecruitWorkerOrder requires non-empty targetTier id',
-      );
-    }
-    final tier = WorkerTier.tryFromId(raw);
-    if (tier == null) {
-      throw ModelValidationException(
-        'RecruitWorkerOrder has unknown targetTier id: "$raw"',
-      );
-    }
-    return RecruitWorkerOrder(targetTier: tier);
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is RecruitWorkerOrder &&
-          runtimeType == other.runtimeType &&
-          targetTier == other.targetTier;
-
-  @override
-  int get hashCode => Object.hash(runtimeType, targetTier);
 }
